# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Trueblue",
        "description": "Trueblue OpenApi description",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/v1/categories/{id}/products": {
            "get": {
                "tags": [
                    "Category"
                ],
                "operationId": "category.listProducts",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/me/reviews/products": {
            "get": {
                "tags": [
                    "Me"
                ],
                "operationId": "me.listReviews",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/v1/products/{id}/reviews": {
            "get": {
                "tags": [
                    "Product"
                ],
                "operationId": "product.listReviews",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Review"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Product"
                ],
                "operationId": "product.createReview",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    },
                                    "rating": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/products/{id}/favourite": {
            "post": {
                "tags": [
                    "Product"
                ],
                "operationId": "product.favouriteProduct.add",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Product"
                ],
                "operationId": "product.favouriteProduct.remove",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/v1/regions/{id}/locations": {
            "get": {
                "tags": [
                    "Region"
                ],
                "operationId": "region.listLocations",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Location"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/v1/regions/{id}": {
            "get": {
                "tags": [
                    "Region"
                ],
                "operationId": "region.showRegion",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Region"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Location": {
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "erBEOm6vAy"
                    },
                    "name": {
                        "type": "string",
                        "example": "銅鑼灣廣場一期"
                    },
                    "lat": {
                        "type": "number",
                        "format": "double",
                        "example": 22.280212
                    },
                    "lng": {
                        "type": "number",
                        "format": "double",
                        "example": 114.180363
                    },
                    "isOpening": {
                        "type": "boolean",
                        "example": false
                    },
                    "openingHours": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "weekday": {
                                    "type": "integer",
                                    "example": 0
                                },
                                "openAt": {
                                    "type": "string",
                                    "example": "08:00"
                                },
                                "closeAt": {
                                    "type": "string",
                                    "example": "17:00"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
            "Product": {
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "erBEOm6vAy"
                    },
                    "title": {
                        "type": "string",
                        "example": "iPad Mini 4"
                    },
                    "description": {
                        "type": "string",
                        "example": "description"
                    },
                    "price": {
                        "type": "number",
                        "example": "100"
                    },
                    "rating": {
                        "type": "number",
                        "example": "4"
                    },
                    "stock": {
                        "type": "number",
                        "example": "0"
                    },
                    "hasStock": {
                        "type": "boolean",
                        "example": false
                    },
                    "isVip": {
                        "type": "boolean",
                        "example": false
                    },
                    "isVariant": {
                        "type": "boolean",
                        "example": false
                    }
                },
                "type": "object"
            },
...
            "Region": {
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "erBEOm6vAy"
                    },
                    "name": {
                        "type": "string",
                        "example": "香港"
                    },
                    "isDefault": {
                        "type": "boolean",
                        "example": false
                    }
                },
                "type": "object"
            },
            "Review": {
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "erBEOm6vAy"
                    },
                    "user": {
                        "type": "object",
                        "$ref": "#/components/schemas/User"
                    },
                    "title": {
                        "type": "string",
                        "example": "title"
                    },
                    "content": {
                        "type": "string",
                        "example": "good product"
                    },
                    "rating": {
                        "type": "number",
                        "example": "4"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "description": "User",
                "properties": {
                    "id": {
                        "schema": "User",
                        "type": "string",
                        "example": "erBEOm6vAy"
                    },
                    "name": {
                        "schema": "User",
                        "type": "string",
                        "example": "Peter Pan"
                    },
                    "email": {
                        "schema": "User",
                        "type": "string",
                        "example": "me@example.com",
                        "nullable": true
                    },
                    "isVip": {
                        "schema": "User",
                        "type": "boolean"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}